+++ CHANGES.md
@@ -4,7 +4,7 @@ Changelog for openvario-shell
 0.3 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Do not write to xcsoar profile on startup if no changes are made
 
 
 0.2 (2020-04-20)